Save analysis result .png to a .part file and rename it when written
The ESpecScreenAAnalyzer looks for the <shot_id>.png file to appear and
sends the contents back to process_message, but it could read the file
before plt::save() had finished (or even started) writing it, leading to
empty returned data.

This ensures the whole file is written before it is read.